pull

+++ sharding-jdbc-example/sharding-jdbc-example-mybatis/src/main/java/com/dangdang/ddframe/rdb/sharding/example/jdbc/entity/Order.java
@@ -1,12 +1,12 @@
-/**
+/*
  * Copyright 1999-2015 dangdang.com.
  * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -19,17 +19,17 @@ package com.dangdang.ddframe.rdb.sharding.example.jdbc.entity;
 
 public final class Order {
     
-    private int orderId;
+    private long orderId;
     
     private int userId;
     
     private String status;
     
-    public int getOrderId() {
+    public long getOrderId() {
         return orderId;
     }
     
-    public void setOrderId(final int orderId) {
+    public void setOrderId(final long orderId) {
         this.orderId = orderId;
     }
